LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/charset-expected.txt

CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: ReferenceError: Can't find variable: unreachable
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: ReferenceError: Can't find variable: unreachable

Harness Error (FAIL), message = ReferenceError: Can't find variable: unreachable
